+++ src/main/java/coupon/domain/Category.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 public enum Category {
     FASHION,
