+++ src/main/java/coupon/domain/Coupon.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
