+++ src/main/java/coupon/domain/MemberCouponEntity.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
