+++ src/main/java/coupon/domain/Money.java
@@ -1,4 +1,4 @@
-package coupon;
+package coupon.domain;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
