+++ src/main/java/coupon/repository/CouponReader.java
@@ -1,5 +1,6 @@
-package coupon;
+package coupon.repository;
 
+import coupon.domain.Coupon;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
