+++ src/main/java/coupon/repository/CouponRepository.java
@@ -1,5 +1,6 @@
-package coupon;
+package coupon.repository;
 
+import coupon.domain.Coupon;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.UUID;
