+++ src/test/java/coupon/MoneyTest.java
@@ -1,5 +1,6 @@
 package coupon;
 
+import coupon.domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
